Add support for channel pinned/muted status through METADATA
First, this implements support for the metadata-2 specification
[1].

The METADATA implementation is bouncer-only (namely, it does not
interact with any upstream METADATA).

The implementation is specific in that there are only 2 supported
keys: "soju.im/pinned" and "soju.im/muted". The user cannot set
any other keys. Those keys can only be "0" or "1", representing
a boolean.

This enables downstreams to get/set two flags on each "message
target" (which can be either users or channels): pinned and muted.

- "soju.im/pinned" is "1" when the channel should be pinned in
  the client, e.g. shown at the top of the channel list
- "soju.im/muted" is "1" when the channel should be muted, e.g.
  have its notifications on non-highlights hidden

The actual specific client behavior for the keys is not
defined here. This just enables clients to get/set the flags.

Last, we only properly support channel targets at the moment.
User targets will be added in a later patch.

[1]: https://github.com/ircv3/ircv3-specifications/blob/master/extensions/metadata.md